Refactored the code so that it uses int16_t instead of any other form of int/uint

+++ Eter/AirMageZephyraCrow.h
@@ -19,6 +19,6 @@ public:
 	std::string getName() const;
 	std::string getDescription() const;
 
-	bool playMage(Board& board, Color color, int8_t x, int8_t y);
+	bool playMage(Board& board, Color color, int16_t x, int16_t y);
 };
