+++ Eter/PowerGust.h
@@ -16,7 +16,7 @@ public:
 	std::string_view getName() const;
 	std::string_view getDescription() const;
 
-	bool checkPower(Board& board, uint8_t x, uint8_t y);
+	bool checkPower(Board& board, int16_t x, int16_t y);
 	void playPower(Board& board);
 };
